refactor: extract share page logic into useSharePage hook
This commit moves the share page's logic into a custom hook
`useSharePage` to improve code organization and reusability.
