+++ app/share/[shareId]/share-client-page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-import { SavedSchedule, saveSchedule } from '@/lib/services/schedules';
+import { useEffect } from 'react';
+import { SavedSchedule } from '@/lib/services/schedules';
 import { useScheduleManagement } from '@/lib/hooks/use-schedule-management';
 import { WeeklySchedule } from '@/components/schedule/weekly-schedule';
 import {
@@ -13,30 +13,26 @@ import {
 } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Save } from 'lucide-react';
-import { useUser } from '@/lib/hooks/use-auth';
-import { useRouter } from 'next/navigation';
 import { SaveScheduleDialog } from '@/components/schedule/save-schedule-dialog';
-import { useMessage } from '@/lib/hooks/use-message';
 import { SelectedCourseList } from '@/components/saved/selected-course-list';
+import { useSharePage } from '@/lib/hooks/use-share-page';
 
 interface ShareClientPageProps {
   schedule: SavedSchedule;
 }
 
 export function ShareClientPage({ schedule }: ShareClientPageProps) {
-  const {
-    selectedCourses,
-    conflicts,
-    timeSlots,
-    getCourseAtTime,
-    setSelectedCoursesDirectly,
-  } = useScheduleManagement();
+  const { selectedCourses, conflicts, setSelectedCoursesDirectly } =
+    useScheduleManagement();
 
-  const { user, loading } = useUser();
-  const router = useRouter();
-  const { showSuccess, showError } = useMessage();
-  const [showSaveDialog, setShowSaveDialog] = useState(false);
-  const [isSaving, setIsSaving] = useState(false);
+  const {
+    loading,
+    showSaveDialog,
+    isSaving,
+    setShowSaveDialog,
+    handleSaveClick,
+    handleConfirmSave,
+  } = useSharePage(schedule);
 
   useEffect(() => {
     if (schedule && schedule.schedule_data) {
@@ -47,39 +43,6 @@ export function ShareClientPage({ schedule }: ShareClientPageProps) {
   const totalCredits =
     schedule.schedule_data?.reduce((sum, course) => sum + course.credits, 0) ??
     0;
-
-  const handleSaveClick = () => {
-    if (!user) {
-      // Jika pengguna tidak login, arahkan ke halaman login
-      // dengan callbackUrl untuk kembali ke halaman ini
-      router.push(`/auth/login?callbackUrl=/share/${schedule.share_id}`);
-    } else {
-      // Jika pengguna sudah login, tampilkan dialog untuk menyimpan
-      setShowSaveDialog(true);
-    }
-  };
-
-  const handleConfirmSave = async (newScheduleName: string) => {
-    if (!schedule.schedule_data) {
-      showError('Tidak ada data jadwal untuk disimpan.');
-      return;
-    }
-
-    setIsSaving(true);
-    try {
-      await saveSchedule(newScheduleName, schedule.schedule_data);
-      showSuccess('Jadwal berhasil disimpan ke akun Anda!');
-      setShowSaveDialog(false);
-      router.push('/saved');
-    } catch (err) {
-      console.error('Failed to save shared schedule:', err);
-      showError(
-        'Gagal menyimpan jadwal. Mungkin sudah ada jadwal dengan nama yang sama.'
-      );
-    } finally {
-      setIsSaving(false);
-    }
-  };
 
   return (
     <>
@@ -108,8 +71,6 @@ export function ShareClientPage({ schedule }: ShareClientPageProps) {
               <WeeklySchedule
                 courses={selectedCourses}
                 conflicts={conflicts}
-                timeSlots={timeSlots}
-                getCourseAtTime={getCourseAtTime}
                 showActions={false}
               />
             </div>
